backend/requirements/dev.txt
-r base.txt
django-debug-toolbar==5.1.0
django-extensions==4.1
ipython==9.2.0
coverage==7.8.0
factory-boy==3.3.3
pytest-django==4.11.1
factory_boy>=3.2.1